+++ src/core/chart-options/AbstractChartOptions.class.ts
@@ -1,4 +1,6 @@
 import { ChartConfiguration } from './ChartConfiguration.class';
+import { OTranslateService } from 'ontimize-web-ngx';
+import { D3LocaleOptions } from './D3LocaleOptions';
 
 declare var d3: any;
 
@@ -6,13 +8,15 @@ export class AbstractChartOptions {
 
   protected chartConf: ChartConfiguration;
   protected type: string;
+  protected d3Locale: D3LocaleOptions;
 
-  constructor(chartConf: ChartConfiguration) {
+  constructor(chartConf: ChartConfiguration, protected translate: OTranslateService) {
     this.chartConf = chartConf;
+    this.d3Locale = new D3LocaleOptions(this.translate);
     this.init(chartConf);
   }
 
-  public init(chartConf: ChartConfiguration) {
+  public init(chartConf: ChartConfiguration): void {
     if (!chartConf) {
       console.error('Chart configuration not found!');
       return;
@@ -21,14 +25,13 @@ export class AbstractChartOptions {
   }
 
   public getOptions(): Object {
-    let opts = {
+    return {
       chart: this.getChartOptions(),
     };
-    return opts;
   }
 
   protected getChartOptions(): Object {
-    let chart = {
+    return {
       type: this.getChartType(),
       height: this.getChartHeight(),
       x: this.getXValue(),
@@ -36,7 +39,6 @@ export class AbstractChartOptions {
       xAxis: this.getXAxisConfiguration(),
       yAxis: this.getYAxisConfiguration()
     };
-    return chart;
   }
 
   protected getChartType(): string {
@@ -47,68 +49,51 @@ export class AbstractChartOptions {
     return this.chartConf.height ? this.chartConf.height : 450;
   }
 
-  protected getXValue(): any {
-    return function (d) { return d.x; };
+  protected getXValue(): Function {
+    return d => d.x;
   }
 
-  protected getYValue(): any {
-    return function (d) { return d.y; };
+  protected getYValue(): Function {
+    return d => d.y;
   }
 
-  protected isArea(): any {
-    return function (d) { return d.area; };
+  protected isArea(): Function {
+    return d => d.area;
   }
 
   protected getXAxisConfiguration(): Object {
-    let xAxis = {
+    return {
       axisLabel: this.chartConf.xLabel ? this.chartConf.xLabel : ''
     };
-    return xAxis;
   }
 
   protected getYAxisConfiguration(): Object {
-    let yAxis = {
+    return {
       axisLabel: this.chartConf.yLabel ? this.chartConf.yLabel : ''
     };
-    return yAxis;
   }
 
-  protected getTickFormatter(type: string) {
-    let formatter;
-
+  protected getTickFormatter(type: string): any {
     switch (type) {
-      case 'intAgrouped':
-        formatter = function (d) {
-          return d3.format(',d')(d);
-        }
-        break;
-      case 'floatAgrouped':
-        formatter = function (d) {
-          return d3.format(',.02f')(d);
-        }
-        break;
+      case 'intGrouped':
+        return d => d3.format(',d')(d);
+      case 'floatGrouped':
+        return d => d3.format(',.02f')(d);
       case 'int':
-        formatter = function (d) {
-          return d3.format('d')(d);
-        };
-        break;
+        return d => d3.format('d')(d);
       case 'float':
-        formatter = function (d) {
-          return d3.format('.02f')(d);
-        };
-        break;
+        return d => d3.format('.02f')(d);
       case 'time':
-        formatter = function (d) {
-          return d3.time.format('%x')(new Date(d));
-        };
-        break;
+        return d => d3.time.format('%x')(new Date(d));
+      case 'timeDay':
+        return d => d3.time.format('%H:%M:%S')(new Date(d));
+      case 'timeDetail':
+        return d => d3.time.format('%x %H:%M:%S')(new Date(d));
       case 'percentage':
-        formatter = function (d) {
-          return d3.format(".0%")(d);
-        };
-        break;
+        return d => d3.format('.0%')(d);
+      default:
+        return void 0;
     }
-
-    return formatter;
   }
+
 }