+++ src/core/chart-options/PieChartOptions.class.ts
@@ -1,21 +1,17 @@
 import { AbstractChartOptions } from './AbstractChartOptions.class';
-import { ChartConfiguration } from './ChartConfiguration.class';
 import { PieChartConfiguration } from './PieChartConfiguration.class';
 
 export class PieChartOptions extends AbstractChartOptions {
-
-  constructor(chartConf: ChartConfiguration) {
-    super(chartConf);
-  }
 
   protected getChartType(): string {
     return 'pieChart';
   }
 
   protected getChartOptions(): Object {
-    let conf = this.chartConf instanceof PieChartConfiguration ? this.chartConf : new PieChartConfiguration();
+    const conf = this.chartConf instanceof PieChartConfiguration ? this.chartConf : new PieChartConfiguration();
+    const valueFormatter = this.getTickFormatter(conf.valueType) || conf.valueType;
 
-    let chart = {
+    const chart = {
       type: this.getChartType(),
       height: this.getChartHeight(),
       width: conf.width,
@@ -25,14 +21,19 @@ export class PieChartOptions extends AbstractChartOptions {
       labelType: conf.labelType,
       labelThreshold: conf.labelThreshold,
       labelSunbeamLayout: conf.labelSunbeamLayout,
+      labelsOutside: conf.labelsOutside,
       showLegend: conf.showLeyend,
       legendPosition: conf.legendPosition,
       legend: conf.legend.getLegendOptions(),
       cornerRadius: conf.cornerRadius,
       donut: conf.donut,
       donutRatio: conf.donutRatio,
-      pieLabelOutside: conf.pieLabelOutside,
-      donutLabelOutside: conf.donutLabelOutside,
+      color: conf.color,
+      valueFormat: valueFormatter,
+      tooltip: {
+        enabled: conf.showTooltip,
+        valueFormatter: valueFormatter
+      }
     };
     return chart;
   }