+++ src/core/chart-options/ScatterChartOption.class.ts
@@ -1,22 +1,18 @@
 import { AbstractChartOptions } from './AbstractChartOptions.class';
-import { ChartConfiguration } from './ChartConfiguration.class';
 import { ScatterChartConfiguration } from './ScatterChartConfiguration.class';
 
 export class ScatterChartOptions extends AbstractChartOptions {
-  constructor(chartConf: ChartConfiguration) {
-    super(chartConf);
-  }
 
   protected getChartType(): string {
     return 'scatterChart';
   }
 
   protected getChartOptions(): Object {
-    let conf = this.chartConf instanceof ScatterChartConfiguration ? this.chartConf : new ScatterChartConfiguration ;
+    const conf = this.chartConf instanceof ScatterChartConfiguration ? this.chartConf : new ScatterChartConfiguration();
     conf.x1Axis.tickFormat = conf.xDataType ? this.getTickFormatter(conf.xDataType) : null;
     conf.y1Axis.tickFormat = conf.yDataType ? this.getTickFormatter(conf.yDataType) : null;
 
-    let chart = {
+    const chart = {
       type: this.getChartType(),
       x: this.getXValue(),
       y: this.getYValue(),
@@ -33,8 +29,9 @@ export class ScatterChartOptions extends AbstractChartOptions {
       showLegend: conf.showLegend,
       margin: conf.margin.getMarginOptions(),
       clipEdge: conf.clipEdge,
-      color: conf.colors
+      color: conf.color
     };
     return chart;
   }
+
 }