+++ src/core/data-adapters/line-data-adapter.ts
@@ -1,77 +1,78 @@
-import { ChartSeries, ChartDataAdapter } from '../../interfaces';
-import {
-  ChartConfiguration
-} from '../../core';
+import { ChartConfiguration } from '../../core';
+import { ChartDataAdapter, ChartSeries } from '../../interfaces';
 import { LineChartConfiguration } from '../chart-options/LineChartConfiguration.class';
 
 export class LineDataAdapter implements ChartDataAdapter {
 
   protected chartConf: ChartConfiguration;
   protected xAxis: string;
-  protected yAxis: Array<String>;
+  protected yAxis: string[];
 
   constructor(chartConf: ChartConfiguration) {
     if (chartConf) {
       this.chartConf = chartConf;
       this.xAxis = this.chartConf.xAxis ? this.chartConf.xAxis : '';
-      let yAxis = this.chartConf.yAxis;
+      const yAxis = this.chartConf.yAxis;
       this.yAxis = yAxis && yAxis.length ? yAxis : [];
     }
   }
-  adaptResult(data: Array<any>): Array<ChartSeries> {
-    let result: Array<ChartSeries> = [];
-    let config = (this.chartConf as LineChartConfiguration);
+
+  public adaptResult(data: any[]): ChartSeries[] {
+    const result: ChartSeries[] = [];
+    const config = (this.chartConf as LineChartConfiguration);
     if (data && data.length) {
-      let seriesValues = this.processSeriesValues(data);
+      const seriesValues = this.processSeriesValues(data);
       this.yAxis.forEach((axis: string, _index: number) => {
-        let serie: ChartSeries = {
-          'key': axis,
+        const serie: ChartSeries = {
+          key: axis,
           values: []
         };
         if (config.isArea && config.isArea[_index]) {
-          serie['area'] = config.isArea[_index];
+          serie.area = config.isArea[_index];
         }
-        if (config.colors && config.colors[_index]) {
-          serie['color'] = config.colors[_index];
+        if (config.color && config.color[_index]) {
+          serie.color = config.color[_index];
         }
         if (config.strokeWidth && config.strokeWidth[_index]) {
-          serie['strokeWidth'] = config.strokeWidth[_index];
+          serie.strokeWidth = config.strokeWidth[_index];
         }
         if (config.classed && config.classed[_index]) {
-          serie['classed'] = config.classed[_index];
+          serie.classed = config.classed[_index];
         }
         let key = axis;
         if (config.translateService) {
           key = config.translateService.get(key);
         }
-        serie['key'] = key;
+        serie.key = key;
 
         seriesValues[axis].sort((a, b) => (a.x > b.x) ? 1 : (b.x > a.x) ? -1 : 0);
 
-        serie['values'] = seriesValues[axis];
+        serie.values = seriesValues[axis];
         result.push(serie);
       });
     }
     return result;
   }
 
-  processSeriesValues(data: Array<any>): Object {
-    let seriesValues = {};
-    var self = this;
+  public processSeriesValues(data: any[]): Object {
+    const seriesValues = {};
+    const self = this;
     data.forEach((item: any, _index: number) => {
       self.yAxis.forEach((axis: string, _indexAxis: number) => {
         if (seriesValues[axis] === undefined) {
           seriesValues[axis] = [];
         }
-        if (item[axis] === undefined)
+        if (item[axis] === undefined) {
           return;
-        let val = {
-          'x': item[self.xAxis],
-          'y': item[axis]
+        }
+        const val = {
+          x: item[self.xAxis],
+          y: item[axis]
         };
         seriesValues[axis].push(val);
       });
     });
     return seriesValues;
   }
+
 }